requirements.txt
Flask==3.0.3
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
rich==13.7.1
